# openapi.yaml
openapi: 3.0.3
info:
  title:  База данных для деканата OpenAPI спецификация
  version: 1.0.11
servers:
  - url: http://localhost:8080/api/v3/
    description: Dev server
paths:
  /student:
    get:
      tags:
        - students
      summary: Метод получения списка студентов
      operationId: findAll
      responses:
        '200':
          description: Успешный ответ со списком студентов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Students'
        '400':          
          description: Неверное значение статуса
        'default':
          description: Всё нестандартное
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Метод создания студента
      tags:
        - students
      operationId: create
      requestBody:
        required: true
        content:
          application/json:
              schema:
                $ref: '#/components/schemas/Student'
      responses:
        '200':
          description: Успешный ответ с созданным студентом
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'  
        '400':          
          description: Неверное значение статуса
        'default':
          description: Всё нестандартное
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error' 
  /student/{student_id}:
    get:
      summary: Метод получения студента по идентификатору
      tags:
        - students
      operationId: findById
      parameters: 
        - name: student_id
          in: path
          required: true
          description: Идентификатор студента
          schema:
            type: string
          example: 4ad76228-517f-4496-aa6c-eef36d3de35a
      responses:
        '200':
          description: Успешный ответ с одним студентом
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'  
        '400':          
          description: Неверное значение статуса
        'default':
          description: Всё нестандартное
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - students
      summary: Удаление студента
      operationId: delete
      parameters:
        - name: student_id
          in: path
          required: true
          description: Идентификатор студента
          schema:
            type: string
          example: 4ad76228-517f-4496-aa6c-eef36d3de35a
      responses:
        '200':
          description: Успешное удаление
          content:
            application/json: {}
        '400':          
          description: Неверное значение статуса
        'default':
          description: Всё нестандартное
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          
components:
  schemas:
    Student:
      type: object
      required:
        - login
        - firstName
        - middleName
        - age
      properties:
        id:
          type: integer
          format: int64
          example: 01d3ff0d-cc37-4863-adef-72d3b0dbf08a
        login:
          type: string
          example: 'Ivan123'
        firstName:
          type: string
          example: Ivan
        middleName:
          type: string
          example: Ivanovich
        lastName:
          type: string
          example: Ivanov
        age:
          type: integer
          format: int32
          example: 23
      xml:
        name: student
    Subject:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          format: int64
          example: 4ad76228-517f-4496-aa6c-eef36d3de35a
        name:
          type: string
          example: Higher mathematics
      xml:
        name: subject
    Students:
      type: array
      items:
        $ref: '#/components/schemas/Student'
    Error:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string